Code review fixes and improvements
